+++ src/dispatcher/flows/account-flow.ts
@@ -1,0 +1,63 @@
+/**
+ * Account dispatch flow — account-based profile (work, personal) with instance isolation.
+ *
+ * Extracted from src/ccs.ts main() profileInfo.type === 'account' branch.
+ * Uses CLAUDE_CONFIG_DIR for per-profile instance isolation on all platforms.
+ */
+
+import { execClaude } from '../../utils/shell-executor';
+import { maybeWarnAboutResumeLaneMismatch } from '../../auth/resume-lane-warning';
+import { resolveNativeClaudeLaunchArgs } from '../environment-builder';
+import type { ProfileDispatchContext } from '../dispatcher-context';
+
+export async function runAccountFlow(ctx: ProfileDispatchContext): Promise<void> {
+  const {
+    profileInfo,
+    claudeCli,
+    nativeClaudeRemainingArgs,
+    InstanceManager,
+    ProfileRegistry,
+    resolveAccountContextPolicy,
+    isAccountContextMetadata,
+  } = ctx;
+
+  const registry = new ProfileRegistry();
+  const instanceMgr = new InstanceManager();
+  const accountMetadata = isAccountContextMetadata(profileInfo.profile)
+    ? profileInfo.profile
+    : undefined;
+  const isBareProfile =
+    typeof profileInfo.profile === 'object' &&
+    profileInfo.profile !== null &&
+    (profileInfo.profile as { bare?: unknown }).bare === true;
+  const contextPolicy = resolveAccountContextPolicy(accountMetadata);
+
+  // Ensure instance exists (lazy init if needed)
+  const instancePath = await instanceMgr.ensureInstance(profileInfo.name, contextPolicy, {
+    bare: isBareProfile,
+  });
+
+  // Update last_used timestamp (check unified config first, fallback to legacy)
+  if (registry.hasAccountUnified(profileInfo.name)) {
+    registry.touchAccountUnified(profileInfo.name);
+  } else {
+    registry.touchProfile(profileInfo.name);
+  }
+
+  // Execute Claude with instance isolation.
+  // Skip WebSearch hook — account profiles use native server-side WebSearch.
+  // Skip Image Analyzer hook — account profiles have native vision support.
+  const envVars: NodeJS.ProcessEnv = {
+    CLAUDE_CONFIG_DIR: instancePath,
+    CCS_PROFILE_TYPE: 'account',
+    CCS_WEBSEARCH_SKIP: '1',
+    CCS_IMAGE_ANALYSIS_SKIP: '1',
+  };
+  await maybeWarnAboutResumeLaneMismatch(profileInfo.name, instancePath, nativeClaudeRemainingArgs);
+  const launchArgs = resolveNativeClaudeLaunchArgs(
+    nativeClaudeRemainingArgs,
+    'account',
+    instancePath
+  );
+  execClaude(claudeCli, launchArgs, envVars);
+}
