+++ src/dispatcher/flows/cliproxy-flow.ts
@@ -1,0 +1,211 @@
+/**
+ * CLIProxy dispatch flow — OAuth-based profiles (gemini, codex, agy, qwen) or user-defined variants.
+ *
+ * Extracted from src/ccs.ts main() profileInfo.type === 'cliproxy' branch.
+ */
+
+import { expandPath } from '../../utils/helpers';
+import { fail, info } from '../../utils/ui';
+import {
+  execClaudeWithCLIProxy,
+  type CLIProxyProvider,
+  ensureCliproxyService,
+  isAuthenticated,
+} from '../../cliproxy';
+import { getEffectiveEnvVars, getCompositeEnvVars } from '../../cliproxy/config/env-builder';
+import { CLIPROXY_DEFAULT_PORT } from '../../cliproxy/config/port-manager';
+import { ensureWebSearchMcpOrThrow } from '../../utils/websearch-manager';
+import { ensureImageAnalysisMcpOrThrow } from '../../utils/image-analysis';
+import {
+  ensureProfileHooks as ensureImageAnalyzerHooks,
+  removeImageAnalysisProfileHook,
+} from '../../utils/hooks/image-analyzer-profile-hook-injector';
+import { prepareImageAnalysisFallbackHook } from '../../utils/hooks';
+import { resolveDroidProvider, type TargetCredentials } from '../../targets';
+import type { ProfileDispatchContext } from '../dispatcher-context';
+
+export async function runCliproxyFlow(ctx: ProfileDispatchContext): Promise<void> {
+  const {
+    profileInfo,
+    resolvedTarget,
+    claudeCli,
+    targetAdapter,
+    targetRemainingArgs,
+    runtimeReasoningOverride,
+    codexRuntimeConfigOverrides,
+    remainingArgs,
+  } = ctx;
+
+  const imageAnalysisMcpReady =
+    resolvedTarget === 'claude' ? ensureImageAnalysisMcpOrThrow() : true;
+  if (resolvedTarget === 'claude') {
+    ensureWebSearchMcpOrThrow();
+  }
+  const provider = profileInfo.provider || (profileInfo.name as CLIProxyProvider);
+  const expandedCliproxySettingsPath = profileInfo.settingsPath
+    ? expandPath(profileInfo.settingsPath)
+    : undefined;
+  if (resolvedTarget === 'claude') {
+    if (imageAnalysisMcpReady) {
+      removeImageAnalysisProfileHook(profileInfo.name, expandedCliproxySettingsPath);
+    } else {
+      const imageAnalysisFallbackHookReady = prepareImageAnalysisFallbackHook();
+      ensureImageAnalyzerHooks({
+        profileName: profileInfo.name,
+        profileType: profileInfo.type,
+        cliproxyProvider: provider,
+        isComposite: profileInfo.isComposite,
+        settingsPath: expandedCliproxySettingsPath,
+        sharedHookInstalled: imageAnalysisFallbackHookReady,
+      });
+    }
+  }
+  const customSettingsPath = profileInfo.settingsPath; // undefined for hardcoded profiles
+  const variantPort = profileInfo.port; // variant-specific port for isolation
+  const cliproxyPort = variantPort || CLIPROXY_DEFAULT_PORT;
+
+  if (resolvedTarget !== 'claude') {
+    const adapter = targetAdapter;
+    if (!adapter) {
+      console.error(fail(`Target adapter not found for "${resolvedTarget}"`));
+      process.exitCode = 1;
+      return;
+    }
+    if (!adapter.supportsProfileType('cliproxy')) {
+      console.error(fail(`${adapter.displayName} does not support CLIProxy profiles`));
+      process.exitCode = 1;
+      return;
+    }
+
+    // Keep CLIProxy management/auth flags on Claude flow only.
+    const unsupportedCliproxyFlags = [
+      '--auth',
+      '--logout',
+      '--accounts',
+      '--add',
+      '--use',
+      '--config',
+      '--headless',
+      '--paste-callback',
+      '--port-forward',
+      '--nickname',
+      '--kiro-auth-method',
+      '--kiro-idc-start-url',
+      '--kiro-idc-region',
+      '--kiro-idc-flow',
+      '--backend',
+      '--proxy-host',
+      '--proxy-port',
+      '--proxy-protocol',
+      '--proxy-auth-token',
+      '--proxy-timeout',
+      '--local-proxy',
+      '--remote-only',
+      '--no-fallback',
+      '--allow-self-signed',
+      '--1m',
+      '--no-1m',
+    ];
+    const providedUnsupportedFlag = unsupportedCliproxyFlags.find(
+      (flag) =>
+        targetRemainingArgs.includes(flag) ||
+        targetRemainingArgs.some((arg) => arg.startsWith(`${flag}=`))
+    );
+    if (providedUnsupportedFlag) {
+      console.error(
+        fail(
+          `${providedUnsupportedFlag} is only supported when running CLIProxy profiles on Claude target`
+        )
+      );
+      console.error(info(`Run with Claude target: ccs ${profileInfo.name} --target claude ...`));
+      process.exitCode = 1;
+      return;
+    }
+
+    // For Droid execution path, require existing OAuth auth and running local proxy.
+    if (profileInfo.isComposite && profileInfo.compositeTiers) {
+      const compositeProviders = [
+        ...new Set(Object.values(profileInfo.compositeTiers).map((tier) => tier.provider)),
+      ] as CLIProxyProvider[];
+      const missingProvider = compositeProviders.find((p) => !isAuthenticated(p));
+      if (missingProvider) {
+        console.error(fail(`Missing OAuth auth for composite tier provider: ${missingProvider}`));
+        console.error(info(`Authenticate first: ccs ${missingProvider} --auth`));
+        process.exitCode = 1;
+        return;
+      }
+    } else if (!isAuthenticated(provider)) {
+      console.error(fail(`No OAuth authentication found for provider: ${provider}`));
+      console.error(info(`Authenticate first: ccs ${provider} --auth`));
+      process.exitCode = 1;
+      return;
+    }
+
+    const ensureServiceResult = await ensureCliproxyService(
+      cliproxyPort,
+      targetRemainingArgs.includes('--verbose') || targetRemainingArgs.includes('-v')
+    );
+    if (!ensureServiceResult.started) {
+      console.error(fail(ensureServiceResult.error || 'Failed to start local CLIProxy service'));
+      process.exitCode = 1;
+      return;
+    }
+
+    const envVars =
+      profileInfo.isComposite && profileInfo.compositeTiers && profileInfo.compositeDefaultTier
+        ? getCompositeEnvVars(
+            profileInfo.compositeTiers,
+            profileInfo.compositeDefaultTier,
+            cliproxyPort,
+            customSettingsPath
+          )
+        : getEffectiveEnvVars(provider, cliproxyPort, customSettingsPath);
+
+    const creds: TargetCredentials = {
+      profile: profileInfo.name,
+      baseUrl: envVars['ANTHROPIC_BASE_URL'] || '',
+      apiKey: envVars['ANTHROPIC_AUTH_TOKEN'] || '',
+      model: envVars['ANTHROPIC_MODEL'] || undefined,
+      provider: resolveDroidProvider({
+        provider: envVars['CCS_DROID_PROVIDER'] || envVars['DROID_PROVIDER'],
+        baseUrl: envVars['ANTHROPIC_BASE_URL'],
+        model: envVars['ANTHROPIC_MODEL'],
+      }),
+      reasoningOverride: runtimeReasoningOverride,
+      runtimeConfigOverrides: codexRuntimeConfigOverrides,
+      envVars,
+    };
+
+    if (!creds.baseUrl || !creds.apiKey) {
+      console.error(
+        fail(
+          `Missing CLIProxy runtime credentials for ${profileInfo.name} (ANTHROPIC_BASE_URL/AUTH_TOKEN)`
+        )
+      );
+      console.error(
+        info('Reconfigure with: ccs config > CLIProxy, or run ccs <provider> --config')
+      );
+      process.exitCode = 1;
+      return;
+    }
+
+    await adapter.prepareCredentials(creds);
+    const targetArgs = adapter.buildArgs(profileInfo.name, targetRemainingArgs, {
+      creds,
+      profileType: profileInfo.type,
+      binaryInfo: ctx.targetBinaryInfo || undefined,
+    });
+    const targetEnv = adapter.buildEnv(creds, profileInfo.type);
+    adapter.exec(targetArgs, targetEnv, { binaryInfo: ctx.targetBinaryInfo || undefined });
+    return;
+  }
+
+  await execClaudeWithCLIProxy(claudeCli, provider, remainingArgs, {
+    customSettingsPath,
+    port: cliproxyPort,
+    isComposite: profileInfo.isComposite,
+    compositeTiers: profileInfo.compositeTiers,
+    compositeDefaultTier: profileInfo.compositeDefaultTier,
+    profileName: profileInfo.name,
+  });
+}
